refactor(cli): decompose RunCommand into run/ modules
Extract the two remaining cohesive areas out of RunCommand.ts, following
the setup/ decomposition pattern (run/FirstRunOnboarding.ts already exists):

- run/RunProfileResolution.ts — settings/profile-source loading, profile
  stack resolution, remote-cache materialization, launch profile layers,
  and the first-run bootstrap profile.
- run/StateWriteReporting.ts — the exit-time state-write diff, the
  state_persistence prompt flow, always-persist recording, and issue
  formatting.

RunCommand.ts is now a <500-line orchestrator, so the file-wide
`/* eslint-disable max-lines */` header is deleted and the repo's
max-lines 500 / complexity 10 rules pass. `createLaunchProfileLayers` and
`loadProfileSources` are re-exported from RunCommand for existing
importers (LintCommand). No behavior change; pinned tests pass unmodified.

Refs #13
