+++ code/cli/src/cli/commands/RunCommand.ts
@@ -1,10 +1,9 @@
-/* eslint-disable max-lines */
-// Provides the command object for launching selected profiles.
-import { existsSync, mkdirSync } from 'node:fs';
+// Provides the command object for launching selected profiles: orchestrates first-run
+// onboarding, profile resolution, composite profile assembly, agent launch, and the
+// exit-time state-write pass. The cohesive pieces live in the run/ modules.
 import type { watch } from 'node:fs';
 import { homedir } from 'node:os';
-import { dirname, join } from 'node:path';
-import { createInterface } from 'node:readline/promises';
+import { join } from 'node:path';
 
 import type { ChildProcess } from 'node:child_process';
 import chalk from 'chalk';
@@ -13,19 +12,9 @@ import spawn from 'cross-spawn';
 
 import type { AgentAdapter, AgentLaunchPlan } from '../../agents/AgentAdapter.js';
 import { createAgentAdapter } from '../../agents/AgentRegistry.js';
-import {
-  createProfileSourceCachePath,
-  createRemoteRepositoryCachePath,
-  redactProfileSourceUriCredentials,
-  resolveRemoteRepositorySubpath,
-} from '../../profiles/ProfileCache.js';
-import { loadLocalProfileSource } from '../../profiles/ProfileLoader.js';
+import { redactProfileSourceUriCredentials } from '../../profiles/ProfileCache.js';
 import type { LoadedProfile } from '../../profiles/ProfileLoader.js';
-import { createEmptyProfile, type Profile } from '../../profiles/Profile.js';
-import type { ProfileSourceReference } from '../../profiles/ProfileSource.js';
-import { resolveProfile } from '../../profiles/ProfileMerger.js';
-import { emptySettings, type Settings } from '../../settings/Settings.js';
-import { loadSettingsWithCachedRemoteSettings } from '../../settings/SettingsLoader.js';
+import type { Profile } from '../../profiles/Profile.js';
 import { exportSystemPromptIfEnabled } from '../../prompts/SystemPromptExport.js';
 import {
   createCompositeProfileRootDirectory,
@@ -34,17 +23,9 @@ import {
 import { renderCompositeProfileTemplates } from '../../compositeProfile/CompositeProfileTemplate.js';
 import {
   createCompositeProfileStateBaseline,
-  detectCompositeProfileStateWrites,
-  persistCompositeProfileStateWrite,
-  recordProfileStatePersistenceOverride,
   updateCompositeProfileStateBaselinePaths,
 } from '../../compositeProfile/StatePersistence.js';
-import type {
-  CompositeProfileStateBaseline,
-  CompositeProfileStatePath,
-  CompositeProfileStateWriteIssue,
-  CompositeProfileStateWritePrompt,
-} from '../../compositeProfile/StatePersistence.js';
+import type { CompositeProfileStateWritePrompt } from '../../compositeProfile/StatePersistence.js';
 import {
   watchCompositeProfileInputs,
   watchCompositeProfileStateWrites,
@@ -66,12 +47,22 @@ import type { CommandObject } from './CommandObject.js';
 import { preparePiLoginLaunchPlan } from './PiLoginLaunch.js';
 import {
   emitClaudeLoginHintIfNeeded,
-  isInteractiveRunLaunch,
   prepareFirstRunRuntimeOnboarding,
   resolveRunProgressWriter,
   runClaudeFirstRunOnboardingIfNeeded,
   selectRunAgentId,
 } from './run/FirstRunOnboarding.js';
+import {
+  createFirstRunBootstrapProfile,
+  createLaunchProfileLayers,
+  loadResolvedProfile,
+} from './run/RunProfileResolution.js';
+import type { ResolvedRunProfile } from './run/RunProfileResolution.js';
+import {
+  handleCompositeProfileStateWrites,
+  recordAlwaysStatePersistenceChoice,
+  resolveStateWritePrompt,
+} from './run/StateWriteReporting.js';
 import type { SetupCommandDependencies } from './SetupCommand.js';
 
 export interface RunCommandInput {
@@ -96,6 +87,7 @@ export interface RunCommandResult {
 
 export type { AgentProcessLauncher } from '../../agents/AgentLaunch.js';
 export { resolveAgentLaunchExecutable } from '../../agents/AgentLaunch.js';
+export { createLaunchProfileLayers, loadProfileSources } from './run/RunProfileResolution.js';
 
 export interface RunCommandDependencies extends SetupCommandDependencies {
   readonly adapter?: AgentAdapter;
@@ -322,18 +314,6 @@ const configureRunCommander = (
 
 /* v8 ignore stop */
 
-interface ResolvedRunProfile {
-  readonly profile: Profile;
-  readonly profilePaths: readonly string[];
-  readonly profileFolders: readonly string[];
-  readonly homeDirectory: string;
-  readonly cacheDirectory: string;
-  readonly projectDirectory: string;
-  readonly settings: Settings;
-  readonly settingsPaths: readonly string[];
-  readonly profileLayers: readonly LoadedProfile[];
-}
-
 const failStrictOnWarnings = (adapterId: string, warnings: readonly string[], strict: boolean | undefined): void => {
   if (strict === true && warnings.length > 0) {
     throw new Error(`Strict failed for ${adapterId}: ${warnings.join('; ')}`);
@@ -475,375 +455,10 @@ const reportPreviousSessionJournals = (sessionJournalDirectory: string, dependen
     dependencies.writeError ?? console.error,
   );
 
-interface CompositeProfileStateWriteHandlingInput {
-  readonly adapterId: string;
-  readonly rootDirectory: string;
-  readonly statePaths: readonly CompositeProfileStatePath[];
-  readonly stateBaseline: CompositeProfileStateBaseline;
-  readonly prompt?: CompositeProfileStateWritePrompt;
-  readonly recordAlwaysChoice: (relativePath: string) => string | undefined;
-  readonly notify: (message: string) => void;
-}
-
-const handleCompositeProfileStateWrites = async (
-  input: CompositeProfileStateWriteHandlingInput,
-): Promise<readonly string[]> => {
-  const warnings: string[] = [];
-
-  for (const issue of detectCompositeProfileStateWrites(input.rootDirectory, input.statePaths, input.stateBaseline)) {
-    if (issue.strategy === 'error') {
-      throw new Error(formatCompositeProfileStateWriteIssue(input.adapterId, issue));
-    }
-
-    if (issue.strategy === 'prompt') {
-      warnings.push(...(await handlePromptStateWriteIssue(input, issue)));
-      continue;
-    }
-
-    warnings.push(formatCompositeProfileStateWriteIssue(input.adapterId, issue));
-  }
-
-  return warnings;
-};
-
-const handlePromptStateWriteIssue = async (
-  input: CompositeProfileStateWriteHandlingInput,
-  issue: CompositeProfileStateWriteIssue,
-): Promise<readonly string[]> => {
-  if (issue.unknown) {
-    return [
-      formatCompositeProfileStateWriteIssue(input.adapterId, issue),
-      `state_persistence 'prompt' cannot persist undeclared writes; '${issue.relativePath}' was reported instead.`,
-    ];
-  }
-
-  if (input.prompt === undefined) {
-    return [
-      formatCompositeProfileStateWriteIssue(input.adapterId, issue),
-      `state_persistence prompt for '${issue.relativePath}' skipped: non-interactive session.`,
-    ];
-  }
-
-  const statePath = findDeclaredStatePath(input.statePaths, issue.relativePath);
-  const choice = await input.prompt({
-    agentId: input.adapterId,
-    relativePath: issue.relativePath,
-    sourcePath: statePath.sourcePath,
-  });
-
-  return applyPromptStateWriteChoice(input, statePath, choice);
-};
-
-const applyPromptStateWriteChoice = (
-  input: CompositeProfileStateWriteHandlingInput,
-  statePath: CompositeProfileStatePath,
-  choice: 'persist' | 'discard' | 'always',
-): readonly string[] => {
-  if (choice === 'discard') {
-    input.notify(`Discarded ${input.adapterId} state write to '${statePath.relativePath}'.`);
-    return [];
-  }
-
-  try {
-    persistCompositeProfileStateWrite(input.rootDirectory, statePath);
-  } catch (error) {
-    return [`Could not persist state path '${statePath.relativePath}': ${String(error)}`];
-  }
-
-  input.notify(`Persisted ${input.adapterId} state write '${statePath.relativePath}' to ${statePath.sourcePath}.`);
-
-  if (choice === 'always') {
-    const warning = input.recordAlwaysChoice(statePath.relativePath);
-    return warning === undefined ? [] : [warning];
-  }
-
-  return [];
-};
-
-const findDeclaredStatePath = (
-  statePaths: readonly CompositeProfileStatePath[],
-  relativePath: string,
-): CompositeProfileStatePath => {
-  const statePath = statePaths.find((candidate) => candidate.relativePath === relativePath);
-
-  /* v8 ignore next 3 -- declared prompt issues always originate from a declared state path. */
-  if (statePath === undefined) {
-    throw new Error(`State path '${relativePath}' is not declared by the composite profile.`);
-  }
-
-  return statePath;
-};
-
-// The "always" choice is recorded in the selected profile's own YAML file because profiles
-// are the single source of truth for state_persistence policy; a parallel settings-layer
-// override would create a second precedence system that adapter validation cannot see.
-// Remote/cached profiles are never mutated, so the choice degrades to a one-run persist
-// with an actionable warning.
-const recordAlwaysStatePersistenceChoice = (
-  adapter: AgentAdapter,
-  resolvedProfile: ResolvedRunProfile,
-  relativePath: string,
-): string | undefined => {
-  const declaration = adapter.statePaths?.[relativePath];
-
-  if (declaration === undefined || !declaration.allowedStrategies.includes('symlink')) {
-    return (
-      `Cannot always-persist '${relativePath}': the ${adapter.id} adapter does not allow 'symlink' for it; ` +
-      `the write was persisted once.`
-    );
-  }
-
-  const selectedLayer = [...resolvedProfile.profileLayers]
-    .reverse()
-    .find((layer) => layer.profile.id === resolvedProfile.profile.id);
-
-  if (
-    selectedLayer === undefined ||
-    selectedLayer.source.uri !== undefined ||
-    selectedLayer.source.github !== undefined
-  ) {
-    return (
-      `Cannot record the always-persist choice for '${relativePath}' because profile ` +
-      `'${resolvedProfile.profile.id}' is not a local profile file; the write was persisted once.`
-    );
-  }
-
-  recordProfileStatePersistenceOverride(selectedLayer.profilePath, relativePath, 'symlink');
-  return undefined;
-};
-
-const resolveStateWritePrompt = (
-  dependencies: RunCommandDependencies,
-): CompositeProfileStateWritePrompt | undefined => {
-  if (!isInteractiveRunLaunch(dependencies)) {
-    return undefined;
-  }
-
-  return (
-    dependencies.promptStateWritePersistence ??
-    /* v8 ignore next -- terminal prompting is direct CLI behavior; tests inject a prompt. */
-    createTerminalStateWritePrompt(dependencies.input ?? process.stdin, dependencies.output ?? process.stdout)
-  );
-};
-
-/* v8 ignore start -- readline prompting is direct terminal behavior; tests inject a prompt. */
-const createTerminalStateWritePrompt =
-  (input: NodeJS.ReadableStream, output: NodeJS.WritableStream): CompositeProfileStateWritePrompt =>
-  async (request) => {
-    const readline = createInterface({ input, output });
-
-    try {
-      for (;;) {
-        const answer = (
-          await readline.question(
-            `${request.agentId} wrote state path '${request.relativePath}' (state_persistence 'prompt'). ` +
-              `[p]ersist to ${request.sourcePath ?? 'its durable source'} / [d]iscard / ` +
-              `[a]lways persist for this profile: `,
-          )
-        )
-          .trim()
-          .toLowerCase();
-
-        if (answer === 'p' || answer === 'persist') {
-          return 'persist';
-        }
-
-        if (answer === 'd' || answer === 'discard') {
-          return 'discard';
-        }
-
-        if (answer === 'a' || answer === 'always') {
-          return 'always';
-        }
-      }
-    } finally {
-      readline.close();
-    }
-  };
-/* v8 ignore stop */
-
-const formatCompositeProfileStateWriteIssue = (adapterId: string, issue: CompositeProfileStateWriteIssue): string => {
-  if (issue.unknown) {
-    return `${adapterId} wrote undeclared composite profile state '${issue.relativePath}' and it was not persisted.`;
-  }
-
-  if (issue.strategy === 'symlink') {
-    return `${adapterId} replaced symlinked state path '${issue.relativePath}' and the change was not persisted.`;
-  }
-
-  return `${adapterId} wrote '${issue.relativePath}' with state_persistence '${issue.strategy}' and it was not persisted.`;
-};
-
-const createFirstRunBootstrapProfile = (input: RunCommandInput): ResolvedRunProfile => ({
-  profile: {
-    ...createEmptyProfile('outfitter-bootstrap'),
-    label: 'Outfitter Bootstrap',
-    description: 'Temporary first-run profile that starts Pi before Outfitter settings exist.',
-  },
-  profilePaths: [],
-  profileFolders: [],
-  homeDirectory: input.homeDirectory,
-  cacheDirectory: join(input.homeDirectory, '.outfitter', 'cache'),
-  projectDirectory: input.projectDirectory,
-  settings: emptySettings(),
-  settingsPaths: [],
-  profileLayers: [],
-});
-
-const loadResolvedProfile = (input: RunCommandInput): ResolvedRunProfile => {
-  const loadedSettings = loadSettingsWithCachedRemoteSettings(input);
-
-  if (loadedSettings.issues.length > 0) {
-    throw new Error(`Cannot run with invalid settings: ${loadedSettings.issues.map(formatSettingsIssue).join('; ')}`);
-  }
-
-  ensureConventionalLocalProfileSourceDirectories(loadedSettings.files);
-  const loadedProfiles = loadProfileSources(input.homeDirectory, loadedSettings.settings.profileSources!);
-
-  if (loadedProfiles.issues.length > 0) {
-    throw new Error(`Cannot run with invalid profiles: ${loadedProfiles.issues.map(formatProfileIssue).join('; ')}`);
-  }
-
-  const profileId = selectRunProfileId(input.profileId, loadedSettings.settings.defaultProfile);
-  const resolution = resolveProfile({
-    profiles: loadedProfiles.profiles,
-    profileId,
-  });
-
-  if (resolution.profile === undefined || resolution.issues.length > 0) {
-    throw new Error(`Cannot resolve profile '${profileId}': ${resolution.issues.map(formatProfileIssue).join('; ')}`);
-  }
-
-  const selectedProfile = resolution.profileStack.find((profile) => profile.id === profileId) as Profile;
-
-  if (selectedProfile.template === true) {
-    throw new Error(`Profile '${profileId}' is a template profile and must be inherited by a runnable profile.`);
-  }
-
-  return {
-    profile: resolution.profile,
-    profileLayers: findContributingLoadedProfiles(resolution.profileStack, loadedProfiles.profiles),
-    profilePaths: findContributingProfilePaths(resolution.profileStack, loadedProfiles.profiles),
-    profileFolders: findContributingProfileFolders(resolution.profileStack, loadedProfiles.profiles),
-    homeDirectory: input.homeDirectory,
-    cacheDirectory: loadedSettings.settings.cacheDirectory ?? join(input.homeDirectory, '.outfitter', 'cache'),
-    projectDirectory: input.projectDirectory,
-    settings: loadedSettings.settings,
-    settingsPaths: loadedSettings.files.map((file) => file.location.path),
-  };
-};
-
-const ensureConventionalLocalProfileSourceDirectories = (files: readonly ResolvedRunProfileSettingsFile[]): void => {
-  for (const file of files) {
-    const settingsProfilesPath = join(dirname(file.location.path), 'profiles');
-    const hasConventionalLocalSource = file.settings.profileSources?.some(
-      (source) => source.uri === undefined && source.github === undefined && source.path === settingsProfilesPath,
-    );
-
-    if (hasConventionalLocalSource === true) {
-      mkdirSync(settingsProfilesPath, { recursive: true });
-    }
-  }
-};
-
-type ResolvedRunProfileSettingsFile = ReturnType<typeof loadSettingsWithCachedRemoteSettings>['files'][number];
-
 const withSystemPromptExportPath = (launchPlan: AgentLaunchPlan, outputPath: string | undefined): AgentLaunchPlan =>
   outputPath === undefined
     ? launchPlan
     : { ...launchPlan, env: { ...launchPlan.env, OUTFITTER_SYSTEM_PROMPT_EXPORT_PATH: outputPath } };
-
-const selectRunProfileId = (selectedProfileId: string | undefined, defaultProfileId: string | undefined): string => {
-  if (selectedProfileId !== undefined) {
-    return selectedProfileId;
-  }
-
-  if (defaultProfileId !== undefined) {
-    return defaultProfileId;
-  }
-
-  throw new Error(
-    'Cannot run without a selected profile or default_profile in settings.yml; pass --profile or run `outfitter setup`.',
-  );
-};
-
-const findContributingProfilePaths = (
-  profileStack: readonly Profile[],
-  loadedProfiles: readonly LoadedProfile[],
-): readonly string[] =>
-  findContributingLoadedProfiles(profileStack, loadedProfiles).map((loadedProfile) => loadedProfile.profilePath);
-
-const findContributingProfileFolders = (
-  profileStack: readonly Profile[],
-  loadedProfiles: readonly LoadedProfile[],
-): readonly string[] =>
-  findContributingLoadedProfiles(profileStack, loadedProfiles).flatMap((loadedProfile) =>
-    loadedProfile.resourceRootPath === undefined ? [] : [loadedProfile.resourceRootPath],
-  );
-
-export const createLaunchProfileLayers = (loadedProfiles: readonly LoadedProfile[]) =>
-  loadedProfiles.map((loadedProfile) => ({
-    profile: loadedProfile.profile,
-    profilePath: loadedProfile.profilePath,
-    sourceRootPath: loadedProfile.sourceRootPath,
-    resourceRootPath: loadedProfile.resourceRootPath,
-    layout: loadedProfile.layout,
-  }));
-
-const findContributingLoadedProfiles = (
-  profileStack: readonly Profile[],
-  loadedProfiles: readonly LoadedProfile[],
-): readonly LoadedProfile[] =>
-  profileStack.flatMap((profile) => loadedProfiles.filter((loadedProfile) => loadedProfile.profile.id === profile.id));
-
-export const loadProfileSources = (
-  homeDirectory: string,
-  sources: readonly ProfileSourceReference[],
-): {
-  readonly profiles: readonly LoadedProfile[];
-  readonly issues: readonly { readonly path: string; readonly message: string }[];
-} => {
-  const profiles: LoadedProfile[] = [];
-  const issues: { readonly path: string; readonly message: string }[] = [];
-
-  for (const source of sources) {
-    const materializedSource = materializeSource(homeDirectory, source);
-
-    // A remote source whose cache has never synced (degraded-offline onboarding, blocked
-    // network) contributes no profiles instead of failing the launch; a later successful
-    // `outfitter sync` upgrades it to the full catalog.
-    if (isUnsyncedRemoteProfileSource(source, materializedSource.path)) {
-      continue;
-    }
-
-    const result = loadLocalProfileSource(materializedSource);
-    profiles.push(...result.profiles.map((profile) => ({ ...profile, source })));
-    issues.push(...result.issues);
-  }
-
-  return { profiles, issues };
-};
-
-const isUnsyncedRemoteProfileSource = (source: ProfileSourceReference, materializedPath: string | undefined): boolean =>
-  (source.uri !== undefined || source.github !== undefined) &&
-  materializedPath !== undefined &&
-  !existsSync(materializedPath);
-
-const materializeSource = (homeDirectory: string, source: ProfileSourceReference): ProfileSourceReference => {
-  if (source.uri === undefined && source.github === undefined) {
-    return source;
-  }
-
-  if (source.uri !== undefined && source.ref === undefined && source.path === undefined) {
-    return { path: createProfileSourceCachePath(homeDirectory, source.uri), only: source.only, except: source.except };
-  }
-
-  return {
-    path: resolveRemoteRepositorySubpath(createRemoteRepositoryCachePath(homeDirectory, source), source.path),
-    only: source.only,
-    except: source.except,
-  };
-};
 
 /* v8 ignore start -- the real child-process launcher is direct runtime behavior; tests inject a launcher. */
 const createSpawnLauncher = (): AgentProcessLauncher => ({
@@ -879,12 +494,3 @@ const signalNumbers: Readonly<Partial<Record<NodeJS.Signals, number>>> = {
   SIGINT: 2,
   SIGTERM: 15,
 };
-
-const formatSettingsIssue = (issue: {
-  readonly filePath: string;
-  readonly path: string;
-  readonly message: string;
-}): string => `${issue.filePath}#${issue.path} ${issue.message}`;
-
-const formatProfileIssue = (issue: { readonly path: string; readonly message: string }): string =>
-  `${issue.path} ${issue.message}`;
